// frontend/src/app/modules/client/core/model/staff/staff-detail.model.ts
export class StaffDetailModel {
  employee_id: string;
  first_name: string;
  last_name: string;
  date_of_birth: string;
  dob_bs: string;
  employee_type: string;
  gender: string;
  current_address: string;
  permanent_address: string;
  joining_date: string;
  primary_contact: string;
  secondary_contact: string;
  designation: string;
  qualification: string;
  picture: string;
  blood_group: string;
  nationality: string;
  marital_status: string;
}